+++ package.json
@@ -25,7 +25,8 @@
     "react-p5-wrapper": "^2.0.0",
     "react-scripts": "^3.2.0",
     "react-youtube": "^7.9.0",
-    "rxjs": "^6.5.3"
+    "rxjs": "^6.5.3",
+    "socket.io": "^2.3.0"
   },
   "scripts": {
     "start": "react-scripts --max_old_space_size=4096 start",
