+++ public/index.html
@@ -43,5 +43,6 @@
         }
       });
     </script>
+    <script src="/socket.io/socket.io.js"></script>
   </body>
 </html>
