+++ src/components/PageSwitcher/PageSwitcher.js
@@ -1,70 +1,56 @@
 import React, { useState, useCallback } from "react";
-import { MuseClient } from "muse-js";
-import { Select, Card, Stack, Button, ButtonGroup, Checkbox } from "@shopify/polaris";
+import { MuseClient, zipSamples } from "muse-js";
+import { Card, Stack, Button, ButtonGroup, Checkbox , RangeSlider, Modal, TextContainer,} from "@shopify/polaris";
 import { mockMuseEEG } from "./utils/mockMuseEEG";
 import * as connectionText from "./utils/connectionText";
 import { emptyAuxChannelData } from "./utils/chartOptions";
+import { bandpassFilter, epoch, fft, sliceFFT} from "@neurosity/pipes";
+import { Subject } from "rxjs";
+import { multicast } from "rxjs/operators";
+import { takeUntil } from "rxjs/operators";
+import { timer } from "rxjs";
 
-import * as funRaw from "./components/EEGEduRaw/EEGEduRaw";
-import * as funSpectra from "./components/EEGEduSpectra/EEGEduSpectra";
+import { RenderModule } from "./components/EEGEduSpectra/EEGEduSpectra";
 
-let modules = {
-  "raw": "1. Raw and Filtered Data",
-  "spectra": "2. Frequency Spectra"
-};
-const raw = modules.raw;
-const spectra = modules.spectra;
+const io = require('socket.io-client');
 
-const chartTypes = [
-  { label: raw, value: raw },
-  { label: spectra, value: spectra }
-];
 
 let showAux = true; // if it is even available to press (to prevent in some modules)
 let source = {};
-window.multicasts = {}
-window.subscriptions = {};
+
 
 //-----Setup Constants
-window.multicasts['Raw'] = null;
-window.multicasts['Spectra'] = null;
+let thisMulticast = null;
+let subscription = null;
 
-window.subscriptions['Raw'] = null;
-window.subscriptions['Spectra'] = null;
+function getSettings() {
+  return {
+    cutOffLow: 1,
+    cutOffHigh: 100,
+    interval: 100,
+    bins: 256,
+    sliceFFTLow: 1,
+    sliceFFTHigh: 100,
+    duration: 1024,
+    srate: 256,
+    name: 'Spectra',
+    secondsToSave: 1000,
+    nchans: 4
+  }
+};
+
 
 
 export function PageSwitcher() {
-
-
-
 
   // connection status
   const [status, setStatus] = useState(connectionText.connect);
 
   // data pulled out of multicast$
-  const [rawData, setRawData] = useState(emptyAuxChannelData);
-  const [spectraData, setSpectraData] = useState(emptyAuxChannelData); 
-
-  // for picking a new module
-  const [selected, setSelected] = useState(raw);
+  const [Data, setData] = useState(emptyAuxChannelData); 
 
   // pipe settings
-  const [rawSettings, setRawSettings] = useState(funRaw.getSettings); 
-  const [spectraSettings, setSpectraSettings] = useState(funSpectra.getSettings); 
-
-  // Whenever settings changed
-  const handleSelectChange = useCallback(value => {
-    setSelected(value);
-
-    console.log("Switching to: " + value);
-
-    if (window.subscriptionRaw) window.subscriptionRaw.unsubscribe();
-    if (window.subscriptionSpectra) window.subscriptionSpectra.unsubscribe();
-
-    buildPipes(value);
-    subscriptionSetup(value);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const [Settings, setSettings] = useState(getSettings); 
 
   // for popup flag when recording
   const [recordPop, setRecordPop] = useState(false);
@@ -74,14 +60,6 @@ export function PageSwitcher() {
   const [checked, setChecked] = useState(false);
   const handleChange = useCallback((newChecked) => setChecked(newChecked), []);
       
-      // const [pipeRaw] = useState();
-      // const [multicastRaw] = useState('0');
-      // const [subscriptionRaw] = useState(); 
-      
-      // const [pipeSpectra] = useState();
-      // const [multicastSpectra] = useState();
-      // const [subscriptionSpectra] = useState(); 
-
   // ---- Manage Auxillary channel
 
   window.enableAux = checked;
@@ -91,37 +69,51 @@ export function PageSwitcher() {
     window.nchans = 4;
   }
 
-  switch (selected) {
-    case raw:
-      showAux = true;
-
-      break
-    case spectra:
-      showAux = true;
-      break
-    default:
-      console.log("Error on showAux");
-  }
+  showAux = true;
 
   //---- Main functions to build and setup called once connect pressed
 
-  function buildPipes(value) {
-    funRaw.buildPipe(source, rawSettings);
-    funSpectra.buildPipe(source, spectraSettings);
+  function buildPipes() {
+    if (subscription) subscription.unsubscribe();
+
+    console.log("Building Multicast for " + Settings.name);
+    // Build Pipe 
+    thisMulticast = zipSamples(source.eegReadings$).pipe(
+      bandpassFilter({ 
+        cutoffFrequencies: [Settings.cutOffLow, Settings.cutOffHigh], 
+        nbChannels: Settings.nchans }),
+      epoch({
+        duration: Settings.duration,
+        interval: Settings.interval,
+        samplingRate: Settings.srate
+      }),
+      fft({ bins: Settings.bins }),
+      sliceFFT([Settings.sliceFFTLow, Settings.sliceFFTHigh])
+    ).pipe(multicast(() => new Subject()));
   }
 
-  function subscriptionSetup(value) {
-    switch (value) {
-      case raw:
-        funRaw.setup(setRawData, rawSettings, rawData);
-        break;
-      case spectra:
-        funSpectra.setup(setSpectraData, spectraSettings, spectraData);
-        break;
-      default:
-        console.log(
-          "Error on handle Subscriptions. Couldn't switch to: " + value
-        );
+  function subscriptionSetup() {
+    console.log("Subscribing to " + Settings.name);
+
+    if (thisMulticast) {
+      thisMulticast.connect();
+      subscription = thisMulticast.subscribe(data => {
+        setData(inData => {
+          Object.values(inData).forEach((channel, index) => {
+            channel.datasets[0].data = data.psd[index];
+            channel.xLabels = data.freqs;
+          });
+
+          return {
+            ch0: inData.ch0,
+            ch1: inData.ch1,
+            ch2: inData.ch2,
+            ch3: inData.ch3,
+            ch4: inData.ch4
+          };
+        });
+      });
+      console.log("Subscribed to " + Settings.name);
     }
   }
 
@@ -129,6 +121,8 @@ export function PageSwitcher() {
 
   async function connect() {
     try {
+
+
       if (window.debugWithMock) {
         // Debug with Mock EEG Data
         setStatus(connectionText.connectingMock);
@@ -155,8 +149,8 @@ export function PageSwitcher() {
       ) {
 
         //Build and Setup
-        buildPipes(selected);
-        subscriptionSetup(selected);
+        buildPipes();
+        subscriptionSetup();
       }
 
     } catch (err) {
@@ -170,48 +164,89 @@ export function PageSwitcher() {
     window.location.reload();
   }
 
-  // Display settings sliders
-  function pipeSettingsDisplay() {
-    switch(selected) {
-      case raw:
-        return (
-          funRaw.renderSliders(setRawData, setRawSettings, status, rawSettings, source)
-        );
-      case spectra:
-        return (
-          funSpectra.renderSliders(setSpectraData, setSpectraSettings, status, spectraSettings, source)
-        );
-      default: console.log('Error rendering settings display');
-    }
-  }
-
   // Show the chart
   function renderModules() {
-    switch (selected) {
-      case raw:
-        return <funRaw.renderModule data={rawData} />;
-      case spectra:
-        return <funSpectra.renderModule data={spectraData} />;
-      default:
-        console.log("Error on renderCharts switch.");
-    }
+    return <RenderModule data={Data} />;
   }
  
   // Show the record button
   function renderRecord() {
-    switch (selected) {
-      case raw: 
-        return (
-          funRaw.renderRecord(recordPopChange, recordPop, status, rawSettings, setRawSettings)
-        )    
-      case spectra:
-        return (
-          funSpectra.renderRecord(recordPopChange, recordPop, status, spectraSettings, setSpectraSettings)
-        )      
-      default:   
-        console.log("Error on renderRecord.");
+    function handleSecondsToSaveRangeSliderChange(value) {
+      setSettings(prevState => ({...prevState, secondsToSave: value}));
     }
+
+    return(
+      <Card title={'Stream ' + Settings.name +' Data'} sectioned>
+        <Stack>
+          <RangeSlider 
+            disabled={status === connectionText.connect} 
+            min={2}
+            max={180}
+            label={'Streaming Length: ' + Settings.secondsToSave + ' Seconds'} 
+            value={Settings.secondsToSave} 
+            onChange={handleSecondsToSaveRangeSliderChange} 
+          />
+          <ButtonGroup>
+            <Button 
+              onClick={() => {
+                saveToCSV(Settings);
+                recordPopChange();
+              }}
+              primary={status !== connectionText.connect}
+              disabled={status === connectionText.connect}
+            > 
+              {'Stream to websocket'}  
+            </Button>
+          </ButtonGroup>
+          <Modal
+            open={recordPop}
+            onClose={recordPopChange}
+            title="Recording Data"
+          >
+            <Modal.Section>
+              <TextContainer>
+                <p>
+                  Your data is currently recording, 
+                  once complete it will be downloaded as a .csv file 
+                  and can be opened with your favorite spreadsheet program. 
+                  Close this window once the download completes.
+                </p>
+              </TextContainer>
+            </Modal.Section>
+          </Modal>
+        </Stack>
+      </Card>
+    )
   }
+
+  async function saveToCSV(Settings) {
+    var socket = await io.connect('75.152.213.182:8080');
+
+    console.log('Streaming ' + Settings.secondsToSave + ' seconds...');
+    var localObservable$ = null;
+
+    // Create timer 
+    const timer$ = timer(Settings.secondsToSave * 1000);
+
+    // put selected observable object into local and start taking samples
+    localObservable$ = thisMulticast.pipe(
+      takeUntil(timer$)
+    );   
+
+    // now with header in place subscribe to each epoch and log it
+    localObservable$.subscribe({
+      next(x) { 
+        // console.log('Next packet: ', x)
+        socket.emit('data', x)
+
+      },
+      error(err) { console.log(err); },
+      complete() { 
+        console.log('Done streaming')
+      }
+    });      
+  }
+
 
   // Render the entire page using above functions
   return (
@@ -255,16 +290,7 @@ export function PageSwitcher() {
           />
         </Stack>
       </Card>
-      <Card title={'Choose your Module'} sectioned>
-        <Select
-          label={""}
-          options={chartTypes}
-          onChange={handleSelectChange}
-          value={selected}
-        />
-      </Card>
       {renderModules()}
-      {pipeSettingsDisplay()}
       {renderRecord()}
     </React.Fragment>
   );
