--- src/components/PageSwitcher/components/EEGEduRaw/translations/en.json
@@ -1,5 +1,0 @@
-{
-  "title": "Raw data",
-  "xlabel": "Time (msec)",
-  "ylabel": "Voltage (\u03BCV)"
-}
